requirements.txt
dingtalk_stream
pyyaml
langchain
